LayoutTests/imported/w3c/web-platform-tests/service-workers/stub-4.3-service-worker-clients-expected.txt


FAIL ServiceWorkerClients interface: existence and properties of interface object assert_own_property: self does not have own property "ServiceWorkerClients" expected property "ServiceWorkerClients" missing
FAIL ServiceWorkerClients interface object length assert_own_property: self does not have own property "ServiceWorkerClients" expected property "ServiceWorkerClients" missing
FAIL ServiceWorkerClients interface object name assert_own_property: self does not have own property "ServiceWorkerClients" expected property "ServiceWorkerClients" missing
FAIL ServiceWorkerClients interface: existence and properties of interface prototype object assert_own_property: self does not have own property "ServiceWorkerClients" expected property "ServiceWorkerClients" missing
FAIL ServiceWorkerClients interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "ServiceWorkerClients" expected property "ServiceWorkerClients" missing
FAIL ServiceWorkerClients interface: operation getServiced() assert_own_property: self does not have own property "ServiceWorkerClients" expected property "ServiceWorkerClients" missing
FAIL ServiceWorkerClients interface: operation reloadAll() assert_own_property: self does not have own property "ServiceWorkerClients" expected property "ServiceWorkerClients" missing
FAIL ServiceWorkerClients must be primary interface of throw new Error ('No object defined for the ServiceWorkerClients interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerClients interface"
FAIL Stringification of throw new Error ('No object defined for the ServiceWorkerClients interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerClients interface"
FAIL ServiceWorkerClients interface: throw new Error ('No object defined for the ServiceWorkerClients interface') must inherit property "getServiced()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerClients interface"
FAIL ServiceWorkerClients interface: throw new Error ('No object defined for the ServiceWorkerClients interface') must inherit property "reloadAll()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerClients interface"
